Move functions
Move start_sniffer_thread(), stop_sniffer_thread()
and related code into sniffer.c.
