+++ boreas/sniffer.c
@@ -29,12 +29,25 @@
 #include <netinet/ip.h>
 #include <stdlib.h>
 #include <string.h>
+#include <unistd.h>
 
 #undef G_LOG_DOMAIN
 /**
  * @brief GLib log domain.
  */
 #define G_LOG_DOMAIN "alive scan"
+
+/* for using int value in #defined string */
+#define STR(X) #X
+#define ASSTR(X) STR (X)
+#define FILTER_STR                                                           \
+  "(ip6 or ip or arp) and (ip6[40]=129 or icmp[icmptype] == icmp-echoreply " \
+  "or dst port " ASSTR (FILTER_PORT) " or arp[6:2]=2)"
+
+/* Conditional variable and mutex to make sure sniffer thread already started
+ * before sending out pings. */
+pthread_mutex_t mutex = PTHREAD_MUTEX_INITIALIZER;
+pthread_cond_t cond = PTHREAD_COND_INITIALIZER;
 
 /**
  * @brief open a new pcap handle ad set provided filter.
@@ -197,4 +210,124 @@ got_packet (u_char *user_data,
           handle_scan_restrictions (scanner, addr_str);
         }
     }
+}
+
+/**
+ * @brief Sniff packets by starting pcap_loop with callback function.
+ *
+ * @param scanner_p Pointer to scanner struct.
+ */
+static void *
+sniffer_thread (void *scanner_p)
+{
+  int ret;
+  struct scanner *scanner = (struct scanner *) scanner_p;
+
+  pthread_mutex_lock (&mutex);
+  pthread_cond_signal (&cond);
+  pthread_mutex_unlock (&mutex);
+
+  /* reads packets until error or pcap_breakloop() */
+  if ((ret =
+         pcap_loop (scanner->pcap_handle, -1, got_packet, (u_char *) scanner))
+      == PCAP_ERROR)
+    g_debug ("%s: pcap_loop error %s", __func__,
+             pcap_geterr (scanner->pcap_handle));
+  else if (ret == 0)
+    g_debug ("%s: count of packets is exhausted", __func__);
+  else if (ret == PCAP_ERROR_BREAK)
+    g_debug ("%s: Loop was successfully broken after call to pcap_breakloop",
+             __func__);
+
+  pthread_exit (0);
+}
+
+/**
+ * @brief Stop the sniffer thread.
+ *
+ * @param scanner Pointer to scanner struct.
+ * @param sniffer_thread_id pthread_t thread id.
+ *
+ * @return 0 on success, other on Error.
+ */
+int
+stop_sniffer_thread (struct scanner *scanner, pthread_t sniffer_thread_id)
+{
+  int err;
+  void *retval;
+
+  g_debug ("%s: Try to stop thread which is sniffing for alive hosts. ",
+           __func__);
+  /* Try to break loop in sniffer thread. */
+  pcap_breakloop (scanner->pcap_handle);
+  /* Give thread chance to exit on its own. */
+  sleep (2);
+
+  /* Cancel thread. May be necessary if pcap_breakloop() does not break the
+   * loop. */
+  err = pthread_cancel (sniffer_thread_id);
+  if (err == ESRCH)
+    g_debug ("%s: pthread_cancel() returned ESRCH; No thread with the "
+             "supplied ID could be found.",
+             __func__);
+
+  /* join sniffer thread*/
+  err = pthread_join (sniffer_thread_id, &retval);
+  if (err == EDEADLK)
+    g_warning ("%s: pthread_join() returned EDEADLK.", __func__);
+  if (err == EINVAL)
+    g_warning ("%s: pthread_join() returned EINVAL.", __func__);
+  if (err == ESRCH)
+    g_warning ("%s: pthread_join() returned ESRCH.", __func__);
+  if (retval == PTHREAD_CANCELED)
+    g_debug ("%s: pthread_join() returned PTHREAD_CANCELED.", __func__);
+
+  g_debug ("%s: Stopped thread which was sniffing for alive hosts.", __func__);
+
+  /* close handle */
+  if (scanner->pcap_handle != NULL)
+    {
+      pcap_close (scanner->pcap_handle);
+    }
+
+  return err;
+}
+
+/**
+ * @brief Start up the sniffer thread.
+ *
+ * @param scanner Pointer to scanner struct.
+ * @param sniffer_thread_id pthread_t thread id.
+ *
+ * @return 0 on success, other on Error.
+ */
+int
+start_sniffer_thread (struct scanner *scanner, pthread_t *sniffer_thread_id)
+{
+  int err;
+
+  scanner->pcap_handle = open_live (NULL, FILTER_STR);
+  if (scanner->pcap_handle == NULL)
+    {
+      g_warning ("%s: Unable to open valid pcap handle.", __func__);
+      return -1;
+    }
+
+  /* Start sniffer thread. */
+  err = pthread_create (sniffer_thread_id, NULL, sniffer_thread, scanner);
+  if (err == EAGAIN)
+    g_warning ("%s: pthread_create() returned EAGAIN: Insufficient resources "
+               "to create thread.",
+               __func__);
+
+  /* Wait for thread to start up before sending out pings. */
+  pthread_mutex_lock (&mutex);
+  pthread_cond_wait (&cond, &mutex);
+  pthread_mutex_unlock (&mutex);
+  /* Mutex and cond not needed anymore. */
+  pthread_mutex_destroy (&mutex);
+  pthread_cond_destroy (&cond);
+  sleep (2);
+
+  return err;
 }
